Userupdate

+++ src/main/java/com/useo/securewebapplication/controller/UserManagementController.java
@@ -2,14 +2,15 @@ package com.useo.securewebapplication.controller;
 
 import com.useo.securewebapplication.MyUserDto;
 import com.useo.securewebapplication.MyUserRepository;
+import com.useo.securewebapplication.UpdateUserDto;
 import com.useo.securewebapplication.model.MyUser;
 
 import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -85,15 +86,18 @@ public class UserManagementController {
     // GET handler to show the update form
     @GetMapping("/update_user")
     public String showUpdateUserForm(@RequestParam("username") String username, Model model) {
-        // Logging fetching user details for update by username
         logger.debug("Fetching user details for update by username: {}", username);
 
         Optional<MyUser> user = userRepository.findByUsername(username);
         if (user.isPresent()) {
-            model.addAttribute("user", user.get());
+            UpdateUserDto updateUserDto = new UpdateUserDto();
+            updateUserDto.setNewUsername(user.get().getUsername());
+            updateUserDto.setEmail(user.get().getEmail());
+
+            model.addAttribute("updateUserDto", updateUserDto);
+            model.addAttribute("oldUsername", username);
             return "update_user";
         } else {
-            // Logging user not found by username
             logger.debug("User not found by username: {}", username);
             model.addAttribute("error", "User not found.");
             return "error_page";
@@ -102,47 +106,49 @@ public class UserManagementController {
 
     @PostMapping("/update_user")
     public String updateUser(@RequestParam("oldUsername") String oldUsername,
-                             @RequestParam("newUsername") String newUsername,
-                             @RequestParam("newPassword") String newPassword,
-                             @RequestParam("email") String email,
+                             @Valid @ModelAttribute("updateUserDto") UpdateUserDto updateUserDto,
+                             BindingResult bindingResult,
                              Model model) {
-        // Logging updating user details for user with old username
         logger.debug("Updating user details for user with old username: {}", oldUsername);
+
+        if (bindingResult.hasErrors()) {
+            logger.debug("Validation errors occurred during user update");
+            model.addAttribute("oldUsername", oldUsername);
+            return "update_user";
+        }
 
         Optional<MyUser> userOptional = userRepository.findByUsername(oldUsername);
         if (userOptional.isPresent()) {
             MyUser existingUser = userOptional.get();
 
-            Optional<MyUser> userWithNewUsername = userRepository.findByUsername(newUsername);
-            Optional<MyUser> userWithNewEmail = userRepository.findByEmail(email);
+            Optional<MyUser> userWithNewUsername = userRepository.findByUsername(updateUserDto.getNewUsername());
+            Optional<MyUser> userWithNewEmail = userRepository.findByEmail(updateUserDto.getEmail());
 
             if ((userWithNewUsername.isPresent() && !userWithNewUsername.get().getId().equals(existingUser.getId())) ||
                     (userWithNewEmail.isPresent() && !userWithNewEmail.get().getId().equals(existingUser.getId()))) {
-                // Logging username or email already exists for another user
                 logger.debug("Username or email already exists for another user");
                 model.addAttribute("error", "Username and email must be unique.");
-                return "registration_error";
+                model.addAttribute("oldUsername", oldUsername);
+                return "update_user";
             }
 
-            existingUser.setUsername(newUsername);
-            existingUser.setPassword(passwordEncoder.encode(newPassword));
+            existingUser.setUsername(updateUserDto.getNewUsername());
+            existingUser.setPassword(passwordEncoder.encode(updateUserDto.getNewPassword()));
 
-            String anonymizedEmail = MaskingUtils.anonymizeEmail(email);
+            String anonymizedEmail = MaskingUtils.anonymizeEmail(updateUserDto.getEmail());
             existingUser.setEmail(HtmlUtils.htmlEscape(anonymizedEmail));
 
             try {
                 userRepository.save(existingUser);
-                // Logging user details updated successfully for username
-                logger.debug("User details updated successfully for username: {}", newUsername);
-                return "redirect:/update_success";
-            } catch (DataIntegrityViolationException e) {
-                // Logging data integrity violation occurred while updating user details
-                logger.debug("Data integrity violation occurred while updating user details");
-                model.addAttribute("error", "Username and email must be unique.");
-                return "registration_error";
+                logger.debug("User details updated successfully for username: {}", updateUserDto.getNewUsername());
+                return "update_success";
+            } catch (Exception e) {
+                logger.debug("Unexpected error occurred while updating user details", e);
+                model.addAttribute("error", "Unexpected error occurred. Please try again.");
+                model.addAttribute("oldUsername", oldUsername);
+                return "update_user";
             }
         } else {
-            // Logging user not found by username
             logger.debug("User not found by username: {}", oldUsername);
             model.addAttribute("error", "User not found.");
             return "error_page";
@@ -163,19 +169,37 @@ public class UserManagementController {
     public String removeUser(@RequestParam("username") String username,
                              @RequestParam("confirmUsername") String confirmUsername,
                              Model model) {
+
         // Logging removing user with username
         logger.debug("Removing user with username: {}", username);
 
         username = HtmlUtils.htmlEscape(username);
         confirmUsername = HtmlUtils.htmlEscape(confirmUsername);
 
+        // Retrieve logged in user's authentication details
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String loggedInUsername = authentication.getName();
+
+        // Check if user is trying to delete themselves and if they have 'ADMIN' role
+        if (username.equals(loggedInUsername) && authentication.getAuthorities().stream()
+                .anyMatch(r -> r.getAuthority().equals("ROLE_ADMIN"))) {
+            // Log the attempt to delete self as admin
+            logger.warn("Admin user {} attempted to delete themselves", loggedInUsername);
+
+            model.addAttribute("error", "Admin cannot delete yourself - contact your manager");
+
+            // Return to error page or show message as per your application flow
+            return "list_users";
+        }
+
+        // If user is present in database, remove the user. else, error.
         if (username.equals(confirmUsername)) {
             Optional<MyUser> user = userRepository.findByUsername(username);
             if (user.isPresent()) {
                 userRepository.delete(user.get());
                 // Logging user removed successfully with username
                 logger.debug("User with username {} removed successfully", username);
-                return "redirect:/delete_success";
+                return "delete_success";
             } else {
                 // Logging user not found by username
                 logger.debug("User not found by username: {}", username);
@@ -209,11 +233,11 @@ public class UserManagementController {
         if (bindingResult.hasErrors()) {
             // Logging validation errors occurred during registration
             logger.debug("Validation errors occurred during registration");
-            return "register"; // return to registration form to display errors
+            return "register"; // return to registration form
         }
 
         try {
-            // Check if username already exists
+            // Check if *username already exists
             MyUser existingUsername = userRepository.findByUsername(userDto.getUsername()).orElse(null);
             if (existingUsername != null) {
                 // Logging username already exists
@@ -222,7 +246,7 @@ public class UserManagementController {
                 return "registration_error";
             }
 
-            // Check if email already exists
+            // Check if *email already exists
             MyUser existingEmail = userRepository.findByEmail(userDto.getEmail()).orElse(null);
             if (existingEmail != null) {
                 // Logging email already exists
@@ -233,12 +257,11 @@ public class UserManagementController {
 
             // Create and save new user
             MyUser newUser = new MyUser();
-            newUser.setUsername(HtmlUtils.htmlEscape(userDto.getUsername())); // Escape HTML to prevent XSS
-            newUser.setPassword(passwordEncoder.encode(HtmlUtils.htmlEscape(userDto.getPassword()))); // Encode password
+            newUser.setUsername(HtmlUtils.htmlEscape(userDto.getUsername())); // Escape HTML to prevent (Cross-Site Scripting)
+            newUser.setPassword(passwordEncoder.encode(HtmlUtils.htmlEscape(userDto.getPassword())));
 
-            // Optional: Anonymize email if needed (assuming MaskingUtils is a utility class you have)
             String anonymizedEmail = MaskingUtils.anonymizeEmail(userDto.getEmail());
-            newUser.setEmail(HtmlUtils.htmlEscape(anonymizedEmail)); // Escape email as well
+            newUser.setEmail(HtmlUtils.htmlEscape(anonymizedEmail)); // Escape HTML to prevent (Cross-Site Scripting)
 
             newUser.setRole("USER");
 
@@ -246,7 +269,7 @@ public class UserManagementController {
 
             // Logging user registered successfully
             logger.debug("User registered successfully: {}", userDto.getUsername());
-            return "redirect:/registration_success";
+            return "registration_success";
         } catch (Exception e) {
             // Logging unexpected error occurred during registration
             logger.debug("Unexpected error occurred during registration", e); // Include exception in log
